pyproject.toml
[tool.poetry]
name = "fastapi-web-tutorial"
version = "0.1.0"
description = ""
authors = ["limstonestone <ehus1118@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_web_tutorial"}]

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
fastapi = "0.70.0"
uvicorn = "0.15.0"
pydantic = {version = "1.9.0", extras = ["email"]}
sqlmodel = "0.0.6"
beanie = "1.13.1"
python-dotenv = "0.20.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
